requirements.txt
django>=1.9
pyyaml
django_markdown2
